requirements.txt
Django==1.6.2
ThePirateBay==1.3.0
cssselect==0.9.1
dateutils==0.6.6
django-filter==0.7
djangorestframework==2.3.13
lxml==3.3.4
purl==0.8
python-dateutil==2.2
pytz==2014.2
six==1.6.1
static==1.0.2
